Refactor: move search/index.js to new pattern
